+++ faynoSync_test.go
@@ -1484,6 +1484,59 @@ func TestPlatformCreateMacos(t *testing.T) {
 	assert.NotEmpty(t, platformIdMacos)
 }
 
+var platformIdMacosTauri string
+
+func TestPlatformCreateMacosTauri(t *testing.T) {
+
+	router := gin.Default()
+	router.Use(utils.AuthMiddleware())
+	w := httptest.NewRecorder()
+
+	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, viper.GetBool("PERFORMANCE_MODE"))
+	router.POST("/platform/create", func(c *gin.Context) {
+		handler.CreatePlatform(c)
+	})
+
+	payload := `{
+		"platform": "macosTauri",
+		"updaters": [
+			{ "type": "manual", "default": false },
+			{ "type": "tauri", "default": true }
+		]
+	}`
+	body := bytes.NewReader([]byte(payload))
+
+	// Create a POST request to the /platform/create endpoint
+	req, err := http.NewRequest("POST", "/platform/create", body)
+	if err != nil {
+		t.Fatal(err)
+	}
+	// Set the Authorization header.
+	req.Header.Set("Authorization", "Bearer "+authToken)
+	// Set the Content-Type header for JSON
+	req.Header.Set("Content-Type", "application/json")
+
+	// Serve the request using the Gin router
+	router.ServeHTTP(w, req)
+	logrus.Infoln("Response Body:", w.Body.String())
+	// Check the response status code (expecting 200 OK)
+	if w.Code != http.StatusOK {
+		t.Fatalf("expected status %d; got %d", http.StatusOK, w.Code)
+	}
+
+	// Parse the JSON response
+	var response map[string]interface{}
+	err = json.Unmarshal(w.Body.Bytes(), &response)
+	if err != nil {
+		t.Fatal(err)
+	}
+	id, idExists := response["createPlatformResult.Created"]
+	assert.True(t, idExists)
+	platformIdMacosTauri = id.(string)
+	assert.True(t, idExists)
+	assert.NotEmpty(t, platformIdMacosTauri)
+}
+
 func TestUpdatePlatformMacos(t *testing.T) {
 	// Initialize Gin router and recorder for the test
 	router := gin.Default()
@@ -1502,7 +1555,8 @@ func TestUpdatePlatformMacos(t *testing.T) {
 		"updaters": [
 			{ "type": "manual", "default": false },
 			{ "type": "squirrel_darwin", "default": true },
-			{ "type": "electron-builder", "default": false }
+			{ "type": "electron-builder", "default": false },
+			{ "type": "tauri", "default": false }
 		]
 	}`, platformIdMacos)
 	body := bytes.NewReader([]byte(payload))
@@ -2167,15 +2221,18 @@ func TestMultipleUploadWithUpdaters(t *testing.T) {
 		Arch        string
 		Updater     string
 		FileName    string
+		Signature   string
 	}{
-		{"updaters", "0.0.1.137", "nightly", false, false, "macosSquirrel", "universalArch", "squirrel_darwin", "testapp.dmg"},
-		{"updaters", "0.0.2.137", "nightly", true, false, "macosSquirrel", "universalArch", "squirrel_darwin", "test.zip"},
-		{"updaters", "0.0.3.137", "nightly", false, false, "macos", "universalArch", "electron-builder", "testapp.dmg"},
-		{"updaters", "0.0.4.137", "nightly", true, false, "macos", "universalArch", "electron-builder", "latest-mac.yml"},
-		{"updaters", "0.0.5.137", "stable", false, true, "windows", "universalArch", "squirrel_windows", "test.exe"},
-		{"updaters", "0.0.6.137", "stable", true, false, "windows", "universalArch", "squirrel_windows", "RELEASES"},
-		{"updaters", "0.0.7.137", "stable", false, false, "windows", "universalArch", "electron-builder", "test.exe"},
-		{"updaters", "0.0.8.137", "stable", true, false, "windows", "universalArch", "electron-builder", "latest.yml"},
+		{"updaters", "0.0.1.137", "nightly", false, false, "macosSquirrel", "universalArch", "squirrel_darwin", "testapp.dmg", ""},
+		{"updaters", "0.0.2.137", "nightly", true, false, "macosSquirrel", "universalArch", "squirrel_darwin", "test.zip", ""},
+		{"updaters", "0.0.1.137", "nightly", false, false, "macosTauri", "universalArch", "tauri", "test.zip", ""},
+		{"updaters", "0.0.2.137", "nightly", true, false, "macosTauri", "universalArch", "tauri", "test.zip", "dW50cnVzdGVkU2lnbmF0dXJlVGhhdExvb2tzTGlrZUFsb25nU3RyaW5n=="},
+		{"updaters", "0.0.3.137", "nightly", false, false, "macos", "universalArch", "electron-builder", "testapp.dmg", ""},
+		{"updaters", "0.0.4.137", "nightly", true, false, "macos", "universalArch", "electron-builder", "latest-mac.yml", ""},
+		{"updaters", "0.0.5.137", "stable", false, true, "windows", "universalArch", "squirrel_windows", "test.exe", ""},
+		{"updaters", "0.0.6.137", "stable", true, false, "windows", "universalArch", "squirrel_windows", "RELEASES", ""},
+		{"updaters", "0.0.7.137", "stable", false, false, "windows", "universalArch", "electron-builder", "test.exe", ""},
+		{"updaters", "0.0.8.137", "stable", true, false, "windows", "universalArch", "electron-builder", "latest.yml", ""},
 	}
 
 	// Iterate through the combinations and upload the file for each combination.
@@ -2205,7 +2262,8 @@ func TestMultipleUploadWithUpdaters(t *testing.T) {
 		if err != nil {
 			t.Fatal(err)
 		}
-		payload := fmt.Sprintf(`{"app_name": "%s", "version": "%s", "channel": "%s", "publish": %v, "critical": %v, "platform": "%s", "arch": "%s", "updater": "%s"}`, combo.AppName, combo.AppVersion, combo.ChannelName, combo.Published, combo.Critical, combo.Platform, combo.Arch, combo.Updater)
+		payload := fmt.Sprintf(`{"app_name": "%s", "version": "%s", "channel": "%s", "publish": %v, "critical": %v, "platform": "%s", "arch": "%s", "updater": "%s", "signature": "%s"}`, combo.AppName, combo.AppVersion, combo.ChannelName, combo.Published, combo.Critical, combo.Platform, combo.Arch, combo.Updater, combo.Signature)
+		logrus.Infoln("Payload: ", payload)
 		_, err = dataPart.Write([]byte(payload))
 		if err != nil {
 			t.Fatal(err)
@@ -2234,7 +2292,8 @@ func TestMultipleUploadWithUpdaters(t *testing.T) {
 		shouldExpectError := (combo.AppVersion == "0.0.1.137" && combo.FileName == "testapp.dmg") ||
 			(combo.AppVersion == "0.0.3.137" && combo.FileName == "testapp.dmg") ||
 			(combo.AppVersion == "0.0.5.137" && combo.FileName == "test.exe") ||
-			(combo.AppVersion == "0.0.7.137" && combo.FileName == "test.exe")
+			(combo.AppVersion == "0.0.7.137" && combo.FileName == "test.exe") ||
+			(combo.AppVersion == "0.0.1.137" && combo.FileName == "test.zip")
 
 		if shouldExpectError {
 			// Expect 400 status code for error cases
@@ -2258,6 +2317,8 @@ func TestMultipleUploadWithUpdaters(t *testing.T) {
 				expectedErrorMsg = "electron-builder updater requires a YML/YAML file for update functionality. Please include a .yml or .yaml file in your upload"
 			case "squirrel_darwin":
 				expectedErrorMsg = "squirrel darwin updater requires a ZIP archive for update functionality. Please include a ZIP file in your upload"
+			case "tauri":
+				expectedErrorMsg = "tauri updater requires a signature parameter for update functionality. Please include a signature in your request"
 			}
 
 			if expectedErrorMsg != "" {
@@ -2344,6 +2405,34 @@ func TestCheckVersionWithUpdaters(t *testing.T) {
 			TestName:     "SquirrelDarwinUpdateAvailable",
 			Owner:        "admin",
 			Updater:      "squirrel_darwin",
+		},
+		{
+			AppName:      "updaters",
+			Version:      "0.0.2.137",
+			ChannelName:  "nightly",
+			ExpectedJSON: map[string]interface{}{"status": "no_content"},
+			ExpectedCode: http.StatusNoContent,
+			Platform:     "macosTauri",
+			Arch:         "universalArch",
+			TestName:     "TauriUpdateNotAvailable",
+			Owner:        "admin",
+			Updater:      "tauri",
+		},
+		{
+			AppName:     "updaters",
+			Version:     "0.0.1.135",
+			ChannelName: "nightly",
+			ExpectedJSON: map[string]interface{}{
+				"signature": "dW50cnVzdGVkU2lnbmF0dXJlVGhhdExvb2tzTGlrZUFsb25nU3RyaW5n==",
+				"version":   "0.0.2.137",
+				"url":       fmt.Sprintf("http://%s/%s/%s", s3Endpoint, s3Bucket, "updaters-admin/nightly/macosTauri/universalArch/updaters-0.0.2.137.zip"),
+			},
+			ExpectedCode: http.StatusOK,
+			Platform:     "macosTauri",
+			Arch:         "universalArch",
+			TestName:     "TauriUpdateAvailable",
+			Owner:        "admin",
+			Updater:      "tauri",
 		},
 		{
 			AppName:      "updaters",
@@ -4268,11 +4357,12 @@ func TestCheckVersion(t *testing.T) {
 			ChannelName: "nightly",
 			ExpectedJSON: map[string]interface{}{
 				// "changelog":        "### Changelog\n",
-				"update_available": false,
-				// "critical":         true,
-				"update_url_dmg": fmt.Sprintf("http://%s/%s/%s", s3Endpoint, s3Bucket, "public%20testapp-admin/nightly/universalPlatform/universalArch/public%20testapp-0.0.1.137.dmg"),
-				"update_url_pkg": fmt.Sprintf("http://%s/%s/%s", s3Endpoint, s3Bucket, "public%20testapp-admin/nightly/universalPlatform/universalArch/public%20testapp-0.0.1.137.pkg"),
-				"update_url":     fmt.Sprintf("http://%s/%s/%s", s3Endpoint, s3Bucket, "public%20testapp-admin/nightly/universalPlatform/universalArch/public%20testapp-0.0.1.137"),
+				"update_available":  false,
+				"critical":          false,
+				"possible_rollback": false,
+				"update_url_dmg":    fmt.Sprintf("http://%s/%s/%s", s3Endpoint, s3Bucket, "public%20testapp-admin/nightly/universalPlatform/universalArch/public%20testapp-0.0.1.137.dmg"),
+				"update_url_pkg":    fmt.Sprintf("http://%s/%s/%s", s3Endpoint, s3Bucket, "public%20testapp-admin/nightly/universalPlatform/universalArch/public%20testapp-0.0.1.137.pkg"),
+				"update_url":        fmt.Sprintf("http://%s/%s/%s", s3Endpoint, s3Bucket, "public%20testapp-admin/nightly/universalPlatform/universalArch/public%20testapp-0.0.1.137"),
 			},
 			ExpectedCode: http.StatusOK,
 			Platform:     "universalPlatform",
@@ -4285,10 +4375,12 @@ func TestCheckVersion(t *testing.T) {
 			Version:     "0.0.2.137",
 			ChannelName: "nightly",
 			ExpectedJSON: map[string]interface{}{
-				"update_available": false,
-				"update_url_dmg":   fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137.dmg"),
-				"update_url_pkg":   fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137.pkg"),
-				"update_url":       fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137"),
+				"update_available":  false,
+				"critical":          false,
+				"possible_rollback": false,
+				"update_url_dmg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137.dmg"),
+				"update_url_pkg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137.pkg"),
+				"update_url":        fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137"),
 			},
 			ExpectedCode: http.StatusOK,
 			Platform:     "universalPlatform",
@@ -4301,9 +4393,15 @@ func TestCheckVersion(t *testing.T) {
 			Version:     "0.0.3.137",
 			ChannelName: "nightly",
 			ExpectedJSON: map[string]interface{}{
-				"error": "requested version 0.0.3.137 is newer than the latest version available",
+				"critical":          true,
+				"possible_rollback": true,
+				"changelog":         "### Changelog\n",
+				"update_available":  false,
+				"update_url_dmg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137.dmg"),
+				"update_url_pkg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137.pkg"),
+				"update_url":        fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.2.137"),
 			},
-			ExpectedCode: http.StatusBadRequest,
+			ExpectedCode: http.StatusOK,
 			Platform:     "universalPlatform",
 			Arch:         "universalArch",
 			TestName:     "NightlyUpdateAvailable",
@@ -4314,10 +4412,12 @@ func TestCheckVersion(t *testing.T) {
 			Version:     "0.0.4.137",
 			ChannelName: "stable",
 			ExpectedJSON: map[string]interface{}{
-				"update_available": false,
-				"update_url_dmg":   fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137.dmg"),
-				"update_url_pkg":   fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137.pkg"),
-				"update_url":       fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137"),
+				"update_available":  false,
+				"critical":          false,
+				"possible_rollback": false,
+				"update_url_dmg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137.dmg"),
+				"update_url_pkg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137.pkg"),
+				"update_url":        fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137"),
 			},
 			ExpectedCode: http.StatusOK,
 			Platform:     "universalPlatform",
@@ -4330,9 +4430,14 @@ func TestCheckVersion(t *testing.T) {
 			Version:     "0.0.5.137",
 			ChannelName: "stable",
 			ExpectedJSON: map[string]interface{}{
-				"error": "requested version 0.0.5.137 is newer than the latest version available",
+				"update_available":  false,
+				"critical":          true,
+				"possible_rollback": true,
+				"update_url_dmg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137.dmg"),
+				"update_url_pkg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137.pkg"),
+				"update_url":        fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fstable%2FuniversalPlatform%2FuniversalArch%2Ftestapp-0.0.4.137"),
 			},
-			ExpectedCode: http.StatusBadRequest,
+			ExpectedCode: http.StatusOK,
 			// Published:    false,
 			Platform: "universalPlatform",
 			Arch:     "universalArch",
@@ -4605,9 +4710,12 @@ func TestCheckVersionWithSameExtensionArtifactsAndDiffPlatformsArchs(t *testing.
 			Version:     "0.0.3.138",
 			ChannelName: "nightly",
 			ExpectedJSON: map[string]interface{}{
-				"error": "requested version 0.0.3.138 is newer than the latest version available",
+				"critical":          false,
+				"possible_rollback": true,
+				"update_available":  false,
+				"update_url_dmg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "testapp-admin%2Fnightly%2FsecondPlatform%2FsecondArch%2Ftestapp-0.0.1.138.dmg"),
 			},
-			ExpectedCode: http.StatusBadRequest,
+			ExpectedCode: http.StatusOK,
 			Platform:     "secondPlatform",
 			Arch:         "secondArch",
 			TestName:     "NightlyUpdateAvailable",
@@ -4703,7 +4811,7 @@ func TestTelemetryWithVariousParams(t *testing.T) {
 				"date_range": dateRange,
 				"admin":      "admin",
 				"summary": map[string]interface{}{
-					"total_requests":               float64(4),
+					"total_requests":               float64(5),
 					"unique_clients":               float64(1),
 					"clients_using_latest_version": float64(0),
 					"clients_outdated":             float64(1),
@@ -4747,7 +4855,7 @@ func TestTelemetryWithVariousParams(t *testing.T) {
 					map[string]interface{}{"date": dailyStats[4], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
 					map[string]interface{}{"date": dailyStats[5], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
 					map[string]interface{}{"date": dailyStats[6], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
-					map[string]interface{}{"date": dailyStats[7], "total_requests": float64(4), "unique_clients": float64(1), "clients_using_latest_version": float64(0), "clients_outdated": float64(1)},
+					map[string]interface{}{"date": dailyStats[7], "total_requests": float64(5), "unique_clients": float64(1), "clients_using_latest_version": float64(0), "clients_outdated": float64(1)},
 				},
 			},
 		},
@@ -4761,7 +4869,7 @@ func TestTelemetryWithVariousParams(t *testing.T) {
 				"date_range": dateRange,
 				"admin":      "admin",
 				"summary": map[string]interface{}{
-					"total_requests":               float64(4),
+					"total_requests":               float64(5),
 					"unique_clients":               float64(1),
 					"clients_using_latest_version": float64(0),
 					"clients_outdated":             float64(1),
@@ -4805,7 +4913,7 @@ func TestTelemetryWithVariousParams(t *testing.T) {
 					map[string]interface{}{"date": dailyStats[4], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
 					map[string]interface{}{"date": dailyStats[5], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
 					map[string]interface{}{"date": dailyStats[6], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
-					map[string]interface{}{"date": dailyStats[7], "total_requests": float64(4), "unique_clients": float64(1), "clients_using_latest_version": float64(0), "clients_outdated": float64(1)},
+					map[string]interface{}{"date": dailyStats[7], "total_requests": float64(5), "unique_clients": float64(1), "clients_using_latest_version": float64(0), "clients_outdated": float64(1)},
 				},
 			},
 		},
@@ -4818,7 +4926,7 @@ func TestTelemetryWithVariousParams(t *testing.T) {
 				"date_range": dateRange,
 				"admin":      "admin",
 				"summary": map[string]interface{}{
-					"total_requests":               float64(4),
+					"total_requests":               float64(5),
 					"unique_clients":               float64(1),
 					"clients_using_latest_version": float64(0),
 					"clients_outdated":             float64(1),
@@ -4862,7 +4970,7 @@ func TestTelemetryWithVariousParams(t *testing.T) {
 					map[string]interface{}{"date": dailyStats[4], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
 					map[string]interface{}{"date": dailyStats[5], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
 					map[string]interface{}{"date": dailyStats[6], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
-					map[string]interface{}{"date": dailyStats[7], "total_requests": float64(4), "unique_clients": float64(1), "clients_using_latest_version": float64(0), "clients_outdated": float64(1)},
+					map[string]interface{}{"date": dailyStats[7], "total_requests": float64(5), "unique_clients": float64(1), "clients_using_latest_version": float64(0), "clients_outdated": float64(1)},
 				},
 			},
 		},
@@ -4875,7 +4983,7 @@ func TestTelemetryWithVariousParams(t *testing.T) {
 				"date_range": dateRange,
 				"admin":      "admin",
 				"summary": map[string]interface{}{
-					"total_requests":               float64(4),
+					"total_requests":               float64(5),
 					"unique_clients":               float64(1),
 					"clients_using_latest_version": float64(0),
 					"clients_outdated":             float64(1),
@@ -4919,7 +5027,7 @@ func TestTelemetryWithVariousParams(t *testing.T) {
 					map[string]interface{}{"date": dailyStats[4], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
 					map[string]interface{}{"date": dailyStats[5], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
 					map[string]interface{}{"date": dailyStats[6], "total_requests": float64(0), "unique_clients": float64(0), "clients_using_latest_version": float64(0), "clients_outdated": float64(0)},
-					map[string]interface{}{"date": dailyStats[7], "total_requests": float64(4), "unique_clients": float64(1), "clients_using_latest_version": float64(0), "clients_outdated": float64(1)},
+					map[string]interface{}{"date": dailyStats[7], "total_requests": float64(5), "unique_clients": float64(1), "clients_using_latest_version": float64(0), "clients_outdated": float64(1)},
 				},
 			},
 		},
@@ -7700,10 +7808,12 @@ func TestCheckVersionWithIntermediate(t *testing.T) {
 			Version:     "0.0.10.138",
 			ChannelName: "nightly",
 			ExpectedJSON: map[string]interface{}{
-				"update_available": false,
-				"update_url_dmg":   fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "newApp-admin%2Fnightly%2FsecondPlatform%2FsecondArch%2FnewApp-0.0.10.138.dmg"),
-				"update_url_pkg":   fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "newApp-admin%2Fnightly%2FsecondPlatform%2FsecondArch%2FnewApp-0.0.10.138.pkg"),
-				"update_url":       fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "newApp-admin%2Fnightly%2FsecondPlatform%2FsecondArch%2FnewApp-0.0.10.138"),
+				"update_available":  false,
+				"critical":          false,
+				"possible_rollback": false,
+				"update_url_dmg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "newApp-admin%2Fnightly%2FsecondPlatform%2FsecondArch%2FnewApp-0.0.10.138.dmg"),
+				"update_url_pkg":    fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "newApp-admin%2Fnightly%2FsecondPlatform%2FsecondArch%2FnewApp-0.0.10.138.pkg"),
+				"update_url":        fmt.Sprintf("%s/%s%s", apiUrl, "download?key=", "newApp-admin%2Fnightly%2FsecondPlatform%2FsecondArch%2FnewApp-0.0.10.138"),
 			},
 			ExpectedCode: http.StatusOK,
 			// Published:    false,
@@ -8106,6 +8216,36 @@ func TestDeletePlatformMacos(t *testing.T) {
 
 	// Create a DELETE request for the /platform/delete endpoint.
 	req, err := http.NewRequest("DELETE", "/platform/delete?id="+platformIdMacos, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Set the Authorization header.
+	req.Header.Set("Authorization", "Bearer "+authToken)
+	// Serve the request using the Gin router.
+	router.ServeHTTP(w, req)
+
+	// Check the response status code.
+	assert.Equal(t, http.StatusOK, w.Code)
+
+	expected := `{"deletePlatformResult.DeletedCount":1}`
+	assert.Equal(t, expected, w.Body.String())
+}
+
+func TestDeletePlatformMacosTauri(t *testing.T) {
+
+	router := gin.Default()
+	router.Use(utils.AuthMiddleware())
+	w := httptest.NewRecorder()
+
+	// Define the route for the /platform/delete endpoint.
+	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, viper.GetBool("PERFORMANCE_MODE"))
+	router.DELETE("/platform/delete", func(c *gin.Context) {
+		handler.DeletePlatform(c)
+	})
+
+	// Create a DELETE request for the /platform/delete endpoint.
+	req, err := http.NewRequest("DELETE", "/platform/delete?id="+platformIdMacosTauri, nil)
 	if err != nil {
 		t.Fatal(err)
 	}