+++ mongod/structs.go
@@ -53,8 +53,9 @@ func NewAppRepository(config *connstring.ConnString, client *mongo.Client) AppRe
 }
 
 type Artifact struct {
-	Link    string
-	Package string
+	Link      string
+	Package   string
+	Signature string
 }
 type Changelog struct {
 	Changes string
@@ -65,6 +66,9 @@ type CheckResult struct {
 	Artifacts              []Artifact
 	Changelog              []Changelog
 	IsRequiredIntermediate bool
+	PossibleRollback       bool
+	LatestVersion          string
+	Signature              string
 }
 
 func (c *appRepository) getBasePipeline() mongo.Pipeline {